+++ src/index.js
@@ -5,6 +5,13 @@ import './css/reset.css';
 import './css/custom.css';
 import store from './store';
 import { Provider } from 'react-redux';
+import { sessionLogin } from './store/member';
+
+// 새로고침시 실행
+if (sessionStorage.getItem('basicInfo')) {
+	const parsed = JSON.parse(sessionStorage.getItem('basicInfo'));
+	store.dispatch(sessionLogin(parsed));
+}
 
 ReactDOM.render(
 	<Provider store={store}>
